refactor: optimized admin category page
